--- x/did/types/errors.go
@@ -1,14 +1,0 @@
-package types
-
-// DONTCOVER
-
-import (
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// x/id module sentinel errors
-var (
-	ErrSample               = sdkerrors.Register(ModuleName, 1100, "sample error")
-	ErrInvalidPacketTimeout = sdkerrors.Register(ModuleName, 1500, "invalid packet timeout")
-	ErrInvalidVersion       = sdkerrors.Register(ModuleName, 1501, "invalid version")
-)
